LA/Baseline-RawNet2/pyproject.toml
[project]
name = "baseline-rawnet2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "librosa==0.7.2",
    "numba==0.48",
    "numpy==1.17.0",
    "protobuf==3.20.3",
    "pyyaml==5.3.1",
    "soundfile>=0.13.1",
    "tensorboard>=2.11.2",
    "tensorboardx==2.1",
    "torch==1.4.0",
    "wandb>=0.16.0",
]
